Cleanup variable usage, particularly for parameters. When invoking an algorithm, a variable syntax is used to refer to the parameter, even though this confuses scope.
WebIDL has a better way of defining parameters and scope, but this is only used for API entrypoints and data-structures. A more comprehensive, but substantially more involved solution would be to define all algorithms as WebIDL methods, and local data-structures using WebIDL so that reference scopes were entirely clear.

Fixes #193.