address #630
I'm not entirely satisfied with the addition in common/terms.html,
as it seems too detailed for the terminology section. But...

...as pointed out in the comment of #630,
the API spec references to this definition in multiple places,
and many of those links are actually for instances of the internal representation,
*not* for instances of the strings or maps that represent term definition in JSON.

So the alternative would be to introduce a new name for those "internal term definition"
and patch all the algorithms... which seems laborious and error-prone.

@gkellogg I don't remember what's the process for synchronizing the files in common with other specs.