fix #633
This reorders and factorizes some parts of the 'Inverse Content Creation' algorithm.
It does not change the behaviour of the algorithm,
except for one corner case:
if the active context has no default language but has a default base direction,
the default combined tag is now _{dir}.
In the previous version, it was @none_{dir}.
I find this more consistent with steps 3.13.4 and 3.15.1,
and I suspect that this former discrepency was actually a bug.